Refactor hive/environment setting
