+++ src/Microsoft.VisualStudio.ProjectSystem.IntegrationTests/TestBase.cs
@@ -1,8 +1,5 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.IO;
-
 using Microsoft.Test.Apex;
 using Microsoft.Test.Apex.VisualStudio;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -14,8 +11,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS
     [TestClass] // AssemblyInitialize won't be found without it
     public abstract class TestBase : VisualStudioHostTest
     {
-        private static string _hiveName;
-
         protected TestBase()
         {
             // TestCleanup will fire up another instance of Visual Studio to reset 
@@ -28,7 +23,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS
 
         protected override VisualStudioHostConfiguration GetHostConfiguration()
         {
-            return new ProjectSystemHostConfiguration(_hiveName);
+            return new ProjectSystemHostConfiguration();
         }
 
         protected override OperationsConfiguration GetOperationsConfiguration()
@@ -41,40 +36,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS
             base.DoHostTestCleanup();
 
             TryShutdownVisualStudioInstance();
-        }
-
-        [AssemblyInitialize]
-        public static void Initialize(TestContext context)
-        {
-            _hiveName = GetVsHiveName(context);
-
-            SetTestInstallationDirectoryIfUnset();
-        }
-
-        private static string GetVsHiveName(TestContext context)
-        {
-            // Get hive from .runsettings if present (command-line)
-            string rootSuffix = (string)context.Properties["VsRootSuffix"];
-            if (!string.IsNullOrEmpty(rootSuffix))
-                return rootSuffix;
-
-            // Otherwise, respect the environment, failing that use the default
-            return Environment.GetEnvironmentVariable("RootSuffix") ?? "Exp";
-        }
-
-        private static void SetTestInstallationDirectoryIfUnset()
-        {
-            string installationUnderTest = "VisualStudio.InstallationUnderTest.Path";
-            if (string.IsNullOrEmpty(Environment.GetEnvironmentVariable(installationUnderTest)))
-            {
-                string vsDirectory = Environment.GetEnvironmentVariable("VSAPPIDDIR");
-                string devenv = Environment.GetEnvironmentVariable("VSAPPIDNAME");
-                if (!string.IsNullOrEmpty(vsDirectory) && !string.IsNullOrEmpty(devenv))
-                {   // Use the same version we're running inside (Test Explorer)
-
-                    Environment.SetEnvironmentVariable(installationUnderTest, Path.Combine(vsDirectory, devenv));
-                }
-            }
         }
     }
 }
